feat: moved tool files to tools folder
