+++ apps/webapp/src/app/app.tsx
@@ -17,7 +17,7 @@ import clsx from "clsx"
 import React, { ChangeEvent, useEffect, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
 
-import { LoadingIconButton } from "../loading-fab/loading-icon-button"
+import { LoadingIconButton } from "../loading-icon-button/loading-icon-button"
 import { RootState } from "../redux/root-reducer"
 import { AppDispatch } from "../redux/store"
 import { TodoState } from "../redux/todos/todos.slice"
